Revert "extending existing powermeter code"
This reverts commit a3a79e82e46437be3e4ad02c45d7c3ef17bd6fca.
